refactor: reimplement server with better interfaces
Lot of experimental ideas in this commit

+++ src/model/position.h
@@ -8,7 +8,7 @@ namespace Model {
 struct Position {
   uint32_t x;
   uint32_t y;
-  uint32_t orientation;
+  uint8_t orientation;
 };
 }
